# storage/api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cipi API Documentation",
        "description": "API Documentation for Cipi Control Panel",
        "version": "1.0"
    },
    "paths": {
        "/files": {
            "get": {
                "tags": [
                    "FileManager"
                ],
                "summary": "List all files and Directory contents",
                "description": "List all all server contents.",
                "operationId": "8d3e347d68cda5dd975eadf31429b9fc",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "pathContents": {
                                                "description": "server content(files and directory)",
                                                "type": "Files and Directories",
                                                "example": "index.txt"
                                            },
                                            "params'": {
                                                "description": "current directory name(s))",
                                                "type": "Directory",
                                                "example": "folder"
                                            },
                                            "path'": {
                                                "description": "File manager root path)",
                                                "type": "Directory",
                                                "example": "folder"
                                            },
                                            " $queryPath'": {
                                                "description": "Specific Server location)",
                                                "type": "Directory",
                                                "example": "folder"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/store": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "files and Directory contents",
                "description": "update or store to a server file.",
                "operationId": "ca392267139fd9ecf0265578e7841c0b",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created request"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/show": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "view files contents",
                "description": "view file manager file in readonly mode.",
                "operationId": "b5b3eac17ff6e9bede63b55b1e1edd35",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/edit": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "Edit file contents",
                "description": "edit file manager file and update.",
                "operationId": "a0dd96a84684befdf65ce547d850b52e",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/delete": {
            "delete": {
                "tags": [
                    "FileManager"
                ],
                "summary": "Edit file contents",
                "description": "delete file",
                "operationId": "d8dd4fceb7bc31f20a3f4b5dacdfb2d0",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/download": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "Download file contents",
                "description": "download file",
                "operationId": "33d806d65a11e36b74f1f82349b2a129",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful content"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/create-directory": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "create a new folder",
                "description": "create a new folder",
                "operationId": "0abdbd98c836c088caf35df52a729b4e",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful content"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/create-file": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "create a new file",
                "description": "create a new file",
                "operationId": "40faf0580389b962d5a7ee402c98dbc9",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created content"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/rename-file": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "Rename a new file",
                "description": "Rename a file",
                "operationId": "c854c1edb93bce3397c8184b4a7f0570",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created content"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/copy-file": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "copy a file to the same or different directory on the server",
                "description": "copy a file to the same or different directory on the server",
                "operationId": "43621587f19ab421e1802881d0af114a",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created content"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/files/move-file": {
            "post": {
                "tags": [
                    "FileManager"
                ],
                "summary": "Move a file to a specific location on the server",
                "description": "create a new file",
                "operationId": "daa990fc2a86e51d98a99a02ebf517b4",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "site-uuid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created content"
                    },
                    "422": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    }
}